Include some more files under C100App namespace
For mutant purposes (and also some additional tests to kill mutants).
