+++ spec/controllers/steps/nature_of_application/case_type_controller_spec.rb
@@ -2,5 +2,5 @@ require 'rails_helper'
 
 RSpec.describe Steps::NatureOfApplication::CaseTypeController, type: :controller do
   it_behaves_like 'a starting point step controller'
-  it_behaves_like 'an intermediate step controller', Steps::NatureOfApplication::CaseTypeForm, CaseTypeDecisionTree
+  it_behaves_like 'an intermediate step controller', Steps::NatureOfApplication::CaseTypeForm, C100App::CaseTypeDecisionTree
 end
